Blog project 4-th stage: - getPostMy method is involved in PostService; - profileMy and profileMyPhoto methods were realized in ApiGeneralController, was fixed upload data without photo; - PreAuthorize annotations were added.

+++ src/main/java/main/api/request/ProfileMyRequest.java
@@ -11,5 +11,5 @@ public class ProfileMyRequest {
     private String name;
     private String email;
     private String password;
-    private int removePhoto;
+    private short removePhoto;
 }
